Completely modular image search
