+++ pages/iconography/[collection]/index.js
@@ -1,14 +1,17 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import { Container, Heading, Text } from '@chakra-ui/react';
+import { Grid, Container, Heading, Text } from '@chakra-ui/react';
 
 import Head from '../../../components/Head';
 import Header from '../../../components/Header';
 import Breadcrumbs from '../../../components/Breadcrumbs';
-import ImageFilter from '../../../components/ImageFilter';
 import ImageList from '../../../components/ImageList';
 import ImageGrid from '../../../components/ImageGrid';
+import Timeline from '../../../components/Timeline';
+import ImageSearch from '../../../components/ImageSearch';
+import ViewButtons from '../../../components/ViewButtons';
+import ImageSort from '../../../components/ImageSort';
 
 import { useImages } from '../../../providers/ImageContext';
 import config from '../../../utils/config';
@@ -31,7 +34,12 @@ const Collection = ({ images, collection }) => {
         <Heading textTransform="capitalize">{collection}</Heading>
       </Container>
       <Container maxW="5xl" pb={5}>
-        <ImageFilter />
+        <Timeline min={1600} max={2020} />
+        <Grid templateColumns="2fr 1fr 1fr" gap="50px" my={5}>
+          <ImageSearch />
+          <ImageSort />
+          <ViewButtons />
+        </Grid>
         <Text>{`${activeImages.length} images found`}</Text>
       </Container>
       {size === 'grid' ? (
